requirements.txt
camelot-py==0.11.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
et-xmlfile==1.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.0
pdfminer.six==20221105
pycparser==2.21
pypdf==3.7.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
PyPDF2==2.12.1
opencv-python==4.7.0.*
# python_version=="3.9.*"
